+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Added some additional functions: `elt`, `butlast`.
 Made more the of the `*suggest*` buffer editable, as users often press
 RET at the end of a heading line.
 
+We can now suggest `(apply #'some-func some-list)` too. Try an input
+of `'(2 3 4)` and an output of `24`.
+
 # v0.1
 
 Initial release.